Added main_nav_graph.xml
Added navigation for all fragment
