+++ app/src/main/java/com/example/internetshop/presentation/activity/fragments/ProductDetailsFragment.kt
@@ -6,14 +6,13 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
-import androidx.core.os.bundleOf
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.viewModels
+import androidx.navigation.fragment.findNavController
 import com.example.internetshop.R
 import com.example.internetshop.data.cache.InternetShopDB
 import com.example.internetshop.databinding.FragmentProductDetailsBinding
 import com.example.internetshop.model.data.di.component.AppComponent
-import com.example.internetshop.presentation.activity.ContainerHolder
 import com.example.internetshop.presentation.viewModel.ProductDetailsViewModel
 import com.squareup.picasso.Picasso
 import javax.inject.Inject
@@ -78,19 +77,10 @@ class ProductDetailsFragment : BaseFragment() {
         viewModel.getProductRx(productId)
 
         reviewButton?.setOnClickListener {
-            (requireActivity() as? ContainerHolder)?.let {
-                val fragment = ReviewFragment().apply {
-                    this.arguments = bundleOf(ReviewFragment.EXTRA_ID_REVIEW to productId)
-                }
-                requireActivity().supportFragmentManager.beginTransaction()
-                    .addToBackStack(null)
-                    .replace(
-                        it.getContainerId()
-                            ?: throw IllegalStateException("Container id must not be null"),
-                        fragment
-                    )
-                    .commit()
-            }
+            val action =
+                ProductDetailsFragmentDirections
+                    .actionProductDetailsFragmentToReviewFragment(productId)
+            findNavController().navigate(action)
         }
 
         viewModel.toastEventLiveData.observe(viewLifecycleOwner, {
