+++ app/src/main/java/com/example/internetshop/presentation/viewModel/ProductsListViewModel.kt
@@ -1,40 +1,60 @@
 package com.example.internetshop.presentation.viewModel
 
+import android.util.Log
 import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-import com.example.internetshop.domain.data.mapper.ProductMapper
+import com.example.internetshop.domain.data.mapper.SimpleProductMapper
 import com.example.internetshop.domain.data.model.product.SimpleProduct
-import com.example.internetshop.domain.data.repository.ProductRepository
+import com.example.internetshop.domain.data.usecase.GetProductsFromCategoryUseCase
+import com.example.internetshop.model.data.viewStates.ProductViewState
 import com.example.internetshop.presentation.utils.SingleLiveEvent
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.schedulers.Schedulers
 import javax.inject.Inject
 
 class ProductsListViewModel @Inject constructor(
-    private val productRepository: ProductRepository,
-    private val productMapper: ProductMapper
-) :
-    ViewModel() {
-    val productsList = MutableLiveData<List<SimpleProduct>>()
-    val openDetailsEvent = SingleLiveEvent<String>()
+    private val fromCategoryUseCase: GetProductsFromCategoryUseCase,
+    private val simpleProductMapper: SimpleProductMapper
+) : BaseViewModel() {
+    val productsList = MutableLiveData<List<ProductViewState>>()
+    val navEventLiveData = SingleLiveEvent<Event>()
 
-    fun onProductClicked(product: SimpleProduct) {
-        openDetailsEvent.value = product.id
+    fun getCategoryProductList(categoryName: String) {
+        compositeDisposable.add(
+            fromCategoryUseCase.execute(categoryName)
+                .subscribeOn(Schedulers.io())
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribe({
+                    productsList.value = it.map {
+                        simpleProductMapper.toProductViewState(it).also {
+                            compositeDisposable.add(it.events.subscribe() {
+                                when (it) {
+                                    is ProductViewState.Event.OnClick -> {
+                                        Log.i(
+                                            "Click event",
+                                            "User clicked on a product: ${it.id}"
+                                        )
+                                        navEventLiveData.value =
+                                            Event.OpenProductDetailEvent(it.id)
+                                    }
+                                }
+                            })
+                        }
+                    }
+                },
+                    {
+                        Log.e("Error", "Error: ${it.message}")
+                    })
+        )
     }
 
-    fun getProductsRx() {
-        productRepository.getProductsRx()
-            .subscribeOn(Schedulers.io())
-            .map {
-                it.map {
-                    productMapper.toSimpleProduct(it)
-                }
-            }
-            .observeOn(AndroidSchedulers.mainThread())
-            .subscribe(
-                {
-                    productsList.value = it
-                }, {
-                })
+    sealed class Event {
+        data class OpenProductDetailEvent(val id: String) : Event()
+        data class ToastEvent(val text: String) : Event()
+    }
+
+    fun map(simpleProduct: SimpleProduct): ProductViewState {
+        simpleProduct.also {
+            return simpleProductMapper.toProductViewState(it)
+        }
     }
 }
