+++ build.gradle
@@ -4,9 +4,11 @@ buildscript {
         google()
         mavenCentral()
     }
+    def nav_version = "2.4.0"
     dependencies {
         classpath "com.android.tools.build:gradle:7.0.3"
         classpath 'org.jetbrains.kotlin:kotlin-gradle-plugin:1.5.31'
+        classpath "androidx.navigation:navigation-safe-args-gradle-plugin:$nav_version"
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
